Fix by comments from PR4
