+++ src/components/constructor-card/constructor-card.js
@@ -4,9 +4,9 @@ import { ConstructorElement, DragIcon } from '@ya.praktikum/react-developer-burg
 import { useDispatch } from 'react-redux';
 import {  deteleIngredient, UPDATE_INGREDIENTS } from '../../services/actions';
 import { useDrag, useDrop } from "react-dnd";
+import PropTypes from 'prop-types';
 
 export default function ConstructorCard({ id, index, moveCard, constructorCard }) {
-
     const dispatch = useDispatch();
 
     const removeIngredient = (card) => {
@@ -75,3 +75,24 @@ export default function ConstructorCard({ id, index, moveCard, constructorCard }
     </li>
   )
 }
+
+ConstructorCard.propTypes = {
+    id: PropTypes.string.isRequired,
+    index: PropTypes.number.isRequired,
+    moveCard: PropTypes.func.isRequired,
+    constructorCard: PropTypes.shape({
+            _id: PropTypes.string.isRequired,
+            name: PropTypes.string.isRequired,
+            type: PropTypes.string.isRequired,
+            proteins: PropTypes.number.isRequired,
+            fat: PropTypes.number.isRequired,
+            carbohydrates: PropTypes.number.isRequired,
+            calories: PropTypes.number.isRequired,
+            price: PropTypes.number.isRequired,
+            image: PropTypes.string.isRequired,
+            image_mobile: PropTypes.string.isRequired,
+            image_large: PropTypes.string.isRequired,
+            __v: PropTypes.number.isRequired,
+            uuid: PropTypes.string.isRequired
+        }).isRequired
+};
