+++ src/components/modal/modal.js
@@ -3,12 +3,8 @@ import styles from './modal.module.css';
 import PropTypes from 'prop-types';
 import { CloseIcon } from '@ya.praktikum/react-developer-burger-ui-components'
 import { ModalOverlay } from '../modal-overlay/modal-overlay'
-import { deleteCartModal, deleteOrderModal } from '../../services/actions';
-import { useDispatch } from 'react-redux';
 
-export default function Modal({children}) {
-
-  const dispatch = useDispatch();
+export default function Modal({children, closeModal}) {
 
   useEffect(() => {
     document.addEventListener('keydown', escButtonHandler)
@@ -17,16 +13,10 @@ export default function Modal({children}) {
     }
   })
 
-  const closeModal = () => {
-    dispatch(deleteCartModal());
-    dispatch(deleteOrderModal());
-  }
-
   const escButtonHandler = (e) => {
     e.preventDefault();
     if(e.key === 'Escape') {
-      dispatch(deleteCartModal());
-      dispatch(deleteOrderModal());
+      closeModal();
     }
   }
 
@@ -45,4 +35,5 @@ export default function Modal({children}) {
 
 Modal.propTypes = {
   children: PropTypes.object.isRequired,
+  closeModal: PropTypes.func.isRequired
 };
